+++ packages/humans/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @abloatai/humans
 
+## 0.51.0
+
+### Patch Changes
+
+- Updated dependencies [3a25ab4]
+  - @abloatai/transaction@0.51.0
+
 ## 0.50.0
 
 ### Minor Changes